LPS-75001 Move LimitedInputStream to com.liferay.portal.kernel.internal.dao.jdbc.
